% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_1D_applyClassInt.R
\name{applyClassInt}
\alias{applyClassInt}
\title{applyClassInt}
\usage{
applyClassInt(x, style = "fisher", breaks = 3)
}
\arguments{
\item{x}{numeric vector.  NA will be removed, but count is included as num.na in the output}

\item{breaks}{number of classes required (details depend on selected style)}

\item{type}{one of  one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks" or "dpih"}
}
\description{
Subroutine for calculating classification intervals (uses the classInt package)
}
\examples{
length.classes <- applyClassInt(SPATData_Samples$LengthMEF,
   style="fisher",breaks=3)
names(length.classes)
length.classes$breaks
length.classes$num.na
length.classes$fit
}
\keyword{breaks}
\keyword{classification}
\keyword{interval,}
