% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_2D_applyKMeans.R
\name{applyKMeans}
\alias{applyKMeans}
\title{applyKMeans}
\usage{
applyKMeans(x, style = "fisher", breaks = 3)
}
\arguments{
\item{x}{numeric vector.  NA will be removed, but count is included as num.na in the output}

\item{breaks}{number of classes required (details depend on selected style)}

\item{type}{one of  one of "fixed", "sd", "equal", "pretty", "quantile", "kmeans", "hclust", "bclust", "fisher", "jenks" or "dpih"}
}
\description{
Subroutine for calculating classification intervals (uses the classInt package)
}
\examples{
petals.classes <- applyClassInt(iris$Petal.Length,
   style="fisher",breaks=3)
petals.classes
}
\keyword{breaks}
\keyword{classification}
\keyword{interval,}
