% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_nD_applyRpart.R
\name{applyRpart}
\alias{applyRpart}
\title{applyRpart}
\usage{
applyRpart(formula, data, minsplit = 20)
}
\arguments{
\item{formula}{a text string of the general structure 
"Response ~ Var1 + Var2". Converted to a formula inside the function}

\item{data}{data frame for the response and predictor variables. predictor variables can be numeric or categorical, and can include missing values.}

\item{minsplit}{smallest number of records for attempting a split. default is 20.}
}
\description{
Subroutine for applying classification trees using the rpart package
}
\examples{
formula.use <- as.formula("Species ~ Sepal.Length + Sepal.Width + Petal.Length + Petal.Width")
test.fit <- applyRpart(formula.use, data = iris,minsplit = 10)
summary(test.fit)
}
\keyword{CART,}
\keyword{and}
\keyword{classification}
\keyword{partitioning}
\keyword{recursive}
\keyword{regression}
\keyword{trees,}
