% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_shiftSeries.R
\name{.offset.fn}
\alias{.offset.fn}
\title{.offset.fn}
\usage{
.offset.fn(x, offset)
}
\arguments{
\item{offset}{offset value}

\item{X}{vector}
}
\description{
This function applies offsets to a single vector
}
\examples{
iris[,1]
.offset.fn(iris[,1],2)
}
\keyword{lag,}
\keyword{lead,}
\keyword{offset}
\keyword{shift,}
