% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotBoxes.R
\name{plotBoxes}
\alias{plotBoxes}
\title{plotBoxes}
\usage{
plotBoxes(
  list.in = NULL,
  y.lab = "Axis Label",
  fill.vec = NULL,
  border.vec = NULL,
  labels = TRUE,
  violin = TRUE,
  y.lim = NULL,
  y.lim.trim = FALSE,
  show.n = TRUE
)
}
\arguments{
\item{list.in}{a list with numerical vectors}

\item{y.lab}{label for the y axis}

\item{fill.vec}{vector with fill colors, one for each element of list.in}

\item{labels}{if TRUE, show median values}

\item{violin}{if TRUE, replace half the box plot with a violin plot}

\item{y.lim}{NULL, or a vector with y axis limits}

\item{y.lim.trim}{if y.lim is NULL, should the automatic y limit be trimmed? (skip for large outliers)}

\item{show.n}{if TRUE, show the sample size for each box}

\item{col.vec}{vector with border colors, one for each element of list.in}
}
\description{
Function to generate side-by-side boxplots with various layout options. Note that
NA values are always dropped.
}
\examples{
plotBoxes(list.in = list(Var1 = rnorm(200,0,1),Var2 = rnorm(200,0.5,3)))
}
\keyword{correlation,}
\keyword{pairwise}
\keyword{retrospective}
