% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotCorrMatrix.R
\name{plotCorrMatrix}
\alias{plotCorrMatrix}
\title{plotCorrMatrix}
\usage{
plotCorrMatrix(X, order = "original", n.groups = NA, label = NULL)
}
\arguments{
\item{X}{a correlation matrix, typically generated by a call to calcCorrMatrix or the base fn cor()}

\item{order}{either "original" or "clustered" (for now, other options to be explored)}

\item{n.groups}{number of groups to mark (applies only if order = "clustered"). NA produces default of round(n/3).}

\item{label}{plot title, default is NULL}
}
\description{
This function generates a plot of a correlation matrix with various options. It is a wrapper function for
the corrplot function from the {corrplot} package.
}
\examples{
M <- cor(mtcars)
plotCorrMatrix(M)
}
\keyword{correlation}
\keyword{matrix,}
\keyword{plot}
