% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotGroup.R
\name{plotGroup}
\alias{plotGroup}
\title{plotGroup}
\usage{
plotGroup(X, agg.idx = NULL, plot.type = "print")
}
\arguments{
\item{X}{a data frame with at least 2 series, the first one is the time step (for now MUST be "yr")}

\item{agg.idx}{function to be used for an aggregate index across series. Default is NULL. Options include any function that works in apply(). Typical examples are "mean", "median". Plan is to include various custom options.}

\item{plot.type}{type of plot to generate: "print", or "shiny" for use in the app. "shiny" type plots use {plotly}}
}
\description{
Function to generate a plot of multiple time series in a group, and optionally include an aggregate index.
}
\examples{
plotGroup(SPATData_EnvCov[,1:5],agg.idx="median",plot.type="print")
}
\keyword{index}
\keyword{line}
\keyword{plot,}
