% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotPair.R
\name{plotPair}
\alias{plotPair}
\title{plotPair}
\usage{
plotPair(X, labels = NULL, layout = "single", style = "default",
  colors = NULL)
}
\arguments{
\item{X}{a data frame with 3 series, the first one is the time step (typically year)}

\item{labels}{vector with labels for the 3 columns of X. if NULL, then col names are used}

\item{layout}{plot layout. currently includes: "single", "2axes", "2panels"}

\item{colors}{vector with colors. Details depend on plot type. If NULL, use a built-in default specific for that plot type.}

\item{layout}{plot style. Currently includes "default"}
}
\description{
Function to plot 2 time series with various options
}
\examples{
plotPair(SPATData_EnvCov[,c("yr","jflow","peak")])
}
\keyword{plot,}
\keyword{series}
\keyword{time}
