% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotRanked.R
\name{plotRanked}
\alias{plotRanked}
\title{plotRanked}
\usage{
plotRanked(data.df, trim = 25, maxvars = NULL, xlim = NULL, flag = NULL)
}
\arguments{
\item{data.df}{a data frame with numerical variables}

\item{trim}{a value between 0 and 50, specifying the percentile to trim on each side of the whiskers (0 = show range from min to max, 10 = show 10th to 90th percentile), Default is 25.}

\item{maxvars}{max number of variables to display, default is NULL (i.e. show all).}

\item{xlim}{a vector with numeric values to bound the x axis}
}
\description{
Function to generate a summary plot with variables ranked by median and showing range.
}
\examples{
plotRanked(SPATData_EnvCov[,2:5])
}
\keyword{range}
\keyword{ranking,}
