% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{plotScatter}
\usage{
plotScatter(X, labels = NULL, colors = "fade", recent.window = 10)
}
\arguments{
\item{X}{a data frame with 3 series (the first one is the time step, typically "yr", used for optional coloring,
and to keep the input object consistent with the other plotting functions)}

\item{labels}{vector with labels for the 2 series. if NULL, then col names are used}

\item{colors}{one of NULL, "fade", "recent" NOT YET IMPLEMENTED}

\item{recent.window}{number to specify how many recent points to highlight. Default is 10 (i.e. last 10yrs)}
}
\description{
Function to generate a scatter plot of 2 time series with a linear regression. NA are filtered out.
}
\examples{
plotScatter(SPATData_EnvCov[,c("EC_jflow","EC_pdo")], labels = NULL, colors = "fade",recent.window = 10)
}
\keyword{plot}
\keyword{plot,}
\keyword{scatter}
\keyword{series,}
\keyword{time}
