% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_plotSPLOM.R
\name{plotSPLOM}
\alias{plotSPLOM}
\title{plotSPLOM}
\usage{
plotSPLOM(data.df)
}
\arguments{
\item{data.df}{a data frame with numerical variables}
}
\description{
Function to generate scatterplot of matrix columns (SPLOM). For now this is using 
the pairs() function from base R.
}
\examples{
plotSPLOM(SPATData_EnvCov[,2:5])
}
\keyword{pairwise}
\keyword{scatterplot}
