% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_shiftSeries.R
\name{shiftSeries}
\alias{shiftSeries}
\title{shiftSeries}
\usage{
shiftSeries(X, offsets = NULL)
}
\arguments{
\item{X}{a data frame with at least 2 numeric columns}

\item{offsets}{vector with offsets for each column in X}
}
\description{
This function applies offsets to the series in a data frame.
}
\examples{
iris.shifted <- shiftSeries(iris,
   offsets =  c(0,2,-1,0,0,0),
head(iris.shifted)
}
\keyword{lag,}
\keyword{lead,}
\keyword{offset}
\keyword{shift,}
