% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPATData_Samples}
\alias{SPATData_Samples}
\title{ASL Data Illustration}
\format{A data frame with  11 variables:
\describe{
  \item{Year}{year of sample}
  \item{Pop}{population sampled}
  \item{Grp_Area}{area grouping for the population}
  \item{Date}{sample date}
  \item{}{}
  ...
}}
\source{
Dummy Data
}
\usage{
SPATData_Samples
}
\description{
A dataset containing age, sex, length samples.
}
\keyword{datasets}
