% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_2D_applyKMeans2D.R
\name{applyKMeans2D}
\alias{applyKMeans2D}
\title{applyKMeans2D}
\usage{
applyKMeans2D(
  x,
  centers = 3,
  iter.max = 100,
  nstart = 5,
  algorithm = "Hartigan-Wong"
)
}
\arguments{
\item{x}{data frame with 2 numeric columns.}

\item{centers}{number of clusters}

\item{algorithm}{one of  "Hartigan-Wong", "Lloyd", "Forgy", "MacQueen"}
}
\description{
Subroutine for applying k-means clustering based on to 2 variables. This is simply a wrapper
   for the kmeans() function from the {stats} package.
}
\examples{
test.clusters <- applyKMeans2D(SPATData_Samples[,c("SampleSW","LengthMEF")])
names(test.clusters)
test.clusters$centers
test.clusters$size
test.clusters$na.num
}
\keyword{clustering}
\keyword{k-means,}
