% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_comPair.R, R/FUNCTION_plotPair.R
\name{comPair}
\alias{comPair}
\title{comPair}
\usage{
comPair(X, order = "original", n.groups = NA, label = NULL)

comPair(X, order = "original", n.groups = NA, label = NULL)
}
\arguments{
\item{X}{a data frame with 2 series}

\item{order}{either "original" or "clustered" (for now, other options to be explored)}

\item{n.groups}{number of groups to mark (applies only if order = "clustered"). NA produces default of round(n/3).}

\item{label}{plot title, default is NULL}

\item{X}{a data frame with 2 series}

\item{order}{either "original" or "clustered" (for now, other options to be explored)}

\item{label}{plot title, default is NULL}

\item{n.groups}{number of groups to mark (applies only if order = "clustered"). NA produces default of round(n/3).}
}
\description{
Function to compute and plot correlations of 2 series (cumulative and by time window)

Function to compute and plot correlations of 2 series (cumulative and by time window)
}
\examples{
M <- cor(mtcars)
plotCorrMatrix(M)
M <- cor(mtcars)
plotCorrMatrix(M)
}
\keyword{correlation}
\keyword{matrix,}
\keyword{plot}
