% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTION_comPair.R
\name{comPair}
\alias{comPair}
\title{comPair}
\usage{
comPair(X, window = 12, plot.type = "print")
}
\arguments{
\item{X}{a data frame with 3 series, the first one is the time step (typically year)}

\item{window}{number of time periods to use for each time window in the retrospective. Default is 12}

\item{plot.type}{type of plot to generate: "none", "print", or "shiny" for use in the app. "shiny" type plots use {plotly}}
}
\description{
Function to compute and plot correlations of 2 series (cumulative and by time window). Wrapper function for the
the runCor() function from the TTR package.
}
\examples{
comPair(SPATData_EnvCov[,c("yr","EC_jflow","EC_pdo")])
}
\keyword{correlation,}
\keyword{pairwise}
\keyword{retrospective}
