% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SUB_2D_plotKMeans2D.R
\name{plotKMeans2D}
\alias{plotKMeans2D}
\title{plotKMeans2D}
\usage{
plotKMeans2D(x, plot.type = "basic", label = "Var Name")
}
\arguments{
\item{x}{object created by applyClassInt()}

\item{plot.type}{one of ""basic", "hist", "dens","cumul"}
}
\description{
Subroutine for creating diagnostic plots of the output from applyClassInt()
}
\examples{
petals.classes <- applyClassInt(iris$Petal.Length,
   style="fisher",breaks=2)
plotClassInt(petals.classes,
   plot.type="cumul",
   label = "Petal Length")
}
\keyword{classification}
\keyword{interval}
