% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SPATData_EnvCov}
\alias{SPATData_EnvCov}
\title{Environmental Covariates Data Illustration}
\format{A data frame with  11 variables:
\describe{
  \item{year}{year of observation}
  \item{aflow}{April flow}
  \item{mflow}{May flow}
  \item{jflow}{June flow}
  \item{}{}
  ...
}}
\source{
Include REF
}
\usage{
SPATData_EnvCov
}
\description{
A dataset containing various environmental covariates used in the 2019 Forecast for 
Fraser River Sockeye Salmon (REF)
}
\keyword{datasets}
